Adding a report runner bin file
The github_report_runner.rb file will run a report given on the command line as
well as pass through any arguments from the command line.

Currently, it will really only handle the sample report in the `reports`
directory.  But, I'll find a way to build the report model outside of the
executable file.